fix: harden managed dolt watchdog parent handling

+++ cmd/gc/dolt_start_managed.go
@@ -48,6 +48,7 @@ const (
 	managedDoltTestModeEnv      = "GC_MANAGED_DOLT_TEST_MODE"
 	managedDoltTestParentPIDEnv = "GC_MANAGED_DOLT_TEST_PARENT_PID"
 	managedDoltTestWatchdogArg  = "__gc-managed-dolt-test-watchdog"
+	// The first ExtraFiles entry is exposed to the child as fd 3.
 	managedDoltTestParentPipeFD = 3
 )
 
@@ -251,41 +252,63 @@ func startManagedDoltSQLServerWithTestWatchdog(cityPath, configFile, logFilePath
 		_ = os.Remove(disarmFile)
 		return managedDoltStartedProcess{}, err
 	}
-	parentPipeRead, parentPipeWrite, err := os.Pipe()
-	if err != nil {
-		_ = os.Remove(disarmFile)
-		return managedDoltStartedProcess{}, fmt.Errorf("create dolt test watchdog parent pipe: %w", err)
+	args := []string{managedDoltTestWatchdogArg, managedDoltTestParentPIDString(), configFile, logFilePath, disarmFile}
+	var parentPipeRead *os.File
+	var parentPipeWrite *os.File
+	if !managedDoltTestHasExternalParent() {
+		parentPipeRead, parentPipeWrite, err = os.Pipe()
+		if err != nil {
+			_ = os.Remove(disarmFile)
+			return managedDoltStartedProcess{}, fmt.Errorf("create dolt test watchdog parent pipe: %w", err)
+		}
+		args = append(args, strconv.Itoa(managedDoltTestParentPipeFD))
 	}
-	cmd := exec.Command(watchdogExecutable, managedDoltTestWatchdogArg, managedDoltTestParentPIDString(), configFile, logFilePath, disarmFile, strconv.Itoa(managedDoltTestParentPipeFD))
+	cmd := exec.Command(watchdogExecutable, args...)
 	cmd.Stderr = logFile
 	cmd.Stdin = nil
 	cmd.Env = doltServerEnv(os.Environ())
-	cmd.ExtraFiles = []*os.File{parentPipeRead}
+	if parentPipeRead != nil {
+		cmd.ExtraFiles = []*os.File{parentPipeRead}
+	}
 	stdout, err := cmd.StdoutPipe()
 	if err != nil {
-		_ = parentPipeRead.Close()
-		_ = parentPipeWrite.Close()
+		if parentPipeRead != nil {
+			_ = parentPipeRead.Close()
+		}
+		if parentPipeWrite != nil {
+			_ = parentPipeWrite.Close()
+		}
 		_ = os.Remove(disarmFile)
 		return managedDoltStartedProcess{}, fmt.Errorf("prepare dolt test watchdog: %w", err)
 	}
 	if err := cmd.Start(); err != nil {
-		_ = parentPipeRead.Close()
-		_ = parentPipeWrite.Close()
+		if parentPipeRead != nil {
+			_ = parentPipeRead.Close()
+		}
+		if parentPipeWrite != nil {
+			_ = parentPipeWrite.Close()
+		}
 		_ = os.Remove(disarmFile)
 		return managedDoltStartedProcess{}, fmt.Errorf("start dolt test watchdog: %w", err)
 	}
-	_ = parentPipeRead.Close()
+	if parentPipeRead != nil {
+		_ = parentPipeRead.Close()
+	}
 	pid, err := readManagedDoltTestWatchdogPID(stdout, cmd.Process.Pid)
 	if err != nil {
 		_ = terminateManagedDoltPID(cityPath, cmd.Process.Pid)
 		_ = cmd.Wait()
-		_ = parentPipeWrite.Close()
+		if parentPipeWrite != nil {
+			_ = parentPipeWrite.Close()
+		}
 		_ = os.Remove(disarmFile)
 		return managedDoltStartedProcess{}, err
 	}
 	go func() {
 		_ = cmd.Wait()
-		_ = parentPipeWrite.Close()
+		if parentPipeWrite != nil {
+			_ = parentPipeWrite.Close()
+		}
 	}()
 	started := managedDoltStartedProcess{
 		CityPath:    cityPath,
@@ -691,6 +714,10 @@ func runManagedDoltTestWatchdog(args []string, stdout, stderr *os.File) int {
 			<-done
 			return 0
 		case <-parentDone:
+			if _, err := os.Stat(disarmFile); err == nil {
+				_ = os.Remove(disarmFile)
+				return 0
+			}
 			_ = terminateManagedDoltPID("", cmd.Process.Pid)
 			<-done
 			return 0
